+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/actuators/actuator_net.py
@@ -160,16 +160,23 @@ class ActuatorNetMLP(DCMotor):
         # compute network inputs
         # -- positions
         pos_input = torch.cat([self._joint_pos_error_history[:, i].unsqueeze(2) for i in self.cfg.input_idx], dim=2)
-        pos_input = pos_input.reshape(self._num_envs * self.num_joints, -1)
+        pos_input = pos_input.view(self._num_envs * self.num_joints, -1)
         # -- velocity
         vel_input = torch.cat([self._joint_vel_history[:, i].unsqueeze(2) for i in self.cfg.input_idx], dim=2)
-        vel_input = vel_input.reshape(self._num_envs * self.num_joints, -1)
+        vel_input = vel_input.view(self._num_envs * self.num_joints, -1)
         # -- scale and concatenate inputs
-        network_input = torch.cat([vel_input * self.cfg.vel_scale, pos_input * self.cfg.pos_scale], dim=1)
+        if self.cfg.input_order == "pos_vel":
+            network_input = torch.cat([pos_input * self.cfg.pos_scale, vel_input * self.cfg.vel_scale], dim=1)
+        elif self.cfg.input_order == "vel_pos":
+            network_input = torch.cat([vel_input * self.cfg.vel_scale, pos_input * self.cfg.pos_scale], dim=1)
+        else:
+            raise ValueError(
+                f"Invalid input order for MLP actuator net: {self.cfg.input_order}. Must be 'pos_vel' or 'vel_pos'."
+            )
 
         # run network inference
-        torques = self.network(network_input).reshape(self._num_envs, self.num_joints)
-        self.computed_effort = torques.reshape(self._num_envs, self.num_joints) * self.cfg.torque_scale
+        torques = self.network(network_input).view(self._num_envs, self.num_joints)
+        self.computed_effort = torques.view(self._num_envs, self.num_joints) * self.cfg.torque_scale
 
         # clip the computed effort based on the motor limits
         self.applied_effort = self._clip_effort(self.computed_effort)